runner: Extract stream module with new function
