+++ core/src/subgraph/mod.rs
@@ -8,6 +8,7 @@ mod metrics;
 mod provider;
 mod registrar;
 mod runner;
+mod stream;
 
 pub use self::instance::SubgraphInstance;
 pub use self::instance_manager::SubgraphInstanceManager;
